lab3/requirements.txt
Pillow==8.3.1
numpy==1.20.3
matplotlib==3.4.2
scipy==1.7.0